+++ packages/bundlekit-bundler-rollup/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @bundlekit/bundler-rollup
 
+## 0.0.5
+
+### Patch Changes
+
+- 9f2f088:
+- 3020202: 修复 bundler-rollup 启动访问是空白
+- Updated dependencies [b8f4151]
+  - @bundlekit/shared-utils@0.0.4
+
 ## 0.0.4
 
 ### Patch Changes