+++ packages/bundlekit-plugin-mock/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bundlekit/plugin-mock
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [b8f4151]
+  - @bundlekit/shared-utils@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes