+++ packages/bundlekit-shared-utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @bundlekit/shared-utils
 
+## 0.0.4
+
+### Patch Changes
+
+- b8f4151: Fix package manager detection bugs:
+  - Fix LRU cache key conflict causing wrong package manager to be detected
+  - Fix async methods being called synchronously in constructor
+  - Fix execa not checking exit code on failure
+  - Fix pnpm add command using 'install' instead of 'add'
+  - Remove incompatible --shamefully-hoist flag
+
 ## 0.0.3
 
 ### Patch Changes